extract feed loader test helpers into a shared protocol extension to remove duplication
